requirements.txt
google-generativeai
langchain_google_genai
sentence-transformers==2.2.2
pinecone-client
langchain==0.0.225
flask
pypdf
python-dotenv

-e .